+++ app/Http/Controllers/Admin/RegistrationsController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Mail;
 use Illuminate\Http\Request;
-use App\User;
+use App\Models\User;
 
 class RegistrationsController extends Controller
 {
