+++ app/Models/EventTrigger.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use App\Http\Controllers\SecretariatController;
 use Carbon\Carbon;
